rampup/requirements.txt
astroid==2.4.2
cffi==1.14.1
cryptography==3.0
Django==1.11.17
django-rest-knox==4.1.0
djangorestframework==3.11.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.5
pycparser==2.20
pylint==2.5.3
pytz==2020.1
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
